+++ src/time_split/_frontend/_plot.py
@@ -1,5 +1,5 @@
 from dataclasses import asdict, dataclass, replace
-from typing import TYPE_CHECKING, Any, Literal
+from typing import TYPE_CHECKING, Any, Literal, Sized
 
 import pandas as pd
 from pandas import Timestamp
@@ -22,6 +22,7 @@ from ..types import (
     Span,
 )
 from ._split import split
+from ._to_string import _PrettyTimestamp
 from ._weight import fold_weight
 
 if TYPE_CHECKING:
@@ -123,16 +124,19 @@ def plot(
     plot_data = _get_plot_data(available, splitter, row_count_bin=row_count_bin, show_removed=show_removed)
 
     if bar_labels is True:
-        bar_labels = settings.DEFAULT_TIME_UNIT if plot_data.available is None else COUNT_ROWS
+        bar_labels = (
+            settings.DEFAULT_TIME_UNIT if (plot_data.available is None or _is_limits(available)) else COUNT_ROWS
+        )
 
     if ax is None:
-        _, ax = plt.subplots(
+        fig, ax = plt.subplots(
             tight_layout=True,
             figsize=(
                 plt.rcParams["figure.figsize"][0],
                 3 + len(plot_data.splits) * 0.5,
             ),
         )
+        fig.autofmt_xdate(ha="center", rotation=15)
 
     _plot_splits(ax, plot_data.splits, removed=plot_data.removed)
 
@@ -169,7 +173,7 @@ def _plot_limits(ax: "Axes", limits: LimitsTuple) -> None:
     from matplotlib.dates import date2num
 
     left, right = limits
-    ax.axvline(left, color="k", ls="--", label="Outer range")
+    ax.axvline(left, color="k", ls="--", label="Available")
     ax.axvline(right, color="k", ls="--")
     left_tick, right_tick = date2num(left), date2num(right)  # type: ignore[no-untyped-call]
     ax.set_xticks([left_tick, *ax.get_xticks(), right_tick])
@@ -351,7 +355,15 @@ def _make_title(available: Any | None, split_kwargs: dict[str, Any]) -> str:
     kwargs = {key: value for key, value in split_kwargs.items() if not is_default(key)}
     if available is None:
         formatted_available = ""
+    elif _is_limits(available):
+        available = sorted(_PrettyTimestamp(a).auto for a in available)
+        available = tuple(available)
+        formatted_available = f", {available=}"  # Probably pre-computed
     else:
         pretty = get_public_module(type(available), resolve_reexport=True, include_name=True)
         formatted_available = f", available={pretty}"
     return f"time_split.split({format_kwargs(kwargs, max_value_length=40)}{formatted_available})"
+
+
+def _is_limits(available: Any) -> bool:
+    return isinstance(available, Sized) and len(available) == 2  # noqa: PLR2004