+++ src/time_split/_frontend/__init__.py
@@ -1,7 +1,7 @@
 from ._plot import plot
 from ._progress import default_metrics_formatter, log_split_progress
 from ._split import split
-from ._to_string import to_string
+from ._to_string import format_expanded_limits, to_string
 from ._weight import fold_weight
 
 __all__ = [
@@ -11,4 +11,5 @@ __all__ = [
     "plot",
     "split",
     "to_string",
+    "format_expanded_limits",
 ]